+++ Realm/Realm/Native/AppConfiguration.cs
@@ -90,5 +90,8 @@ namespace Realms.Sync.Native
         internal UInt64 sync_pong_keep_alive_timeout_ms;
 
         internal UInt64 sync_fast_reconnect_limit;
+
+        [MarshalAs(UnmanagedType.U1)]
+        internal bool use_cache;
     }
 }