+++ Realm/Realm/Schema/RealmSchema.cs
@@ -193,7 +193,7 @@ namespace Realms.Schema
         /// <c>null</c> if <paramref name="objects"/> is <c>null</c>; a <see cref="RealmSchema"/> containing the supplied <see cref="ObjectSchema"/>s otherwise.
         /// </returns>
         [return: NotNullIfNotNull("objects")]
-        public static implicit operator RealmSchema?(ObjectSchema[]? objects) => objects == null ? null : new Builder(objects).Build();
+        public static implicit operator RealmSchema?(ObjectSchema[]? objects) => objects is null ? null : new Builder(objects).Build();
 
         /// <summary>
         /// Constructs a <see cref="RealmSchema"/> from a list of <see cref="ObjectSchema"/> instances.
@@ -206,7 +206,7 @@ namespace Realms.Schema
         /// <c>null</c> if <paramref name="objects"/> is <c>null</c>; a <see cref="RealmSchema"/> containing the supplied <see cref="ObjectSchema"/>s otherwise.
         /// </returns>
         [return: NotNullIfNotNull("objects")]
-        public static implicit operator RealmSchema?(List<ObjectSchema>? objects) => objects == null ? null : new Builder(objects).Build();
+        public static implicit operator RealmSchema?(List<ObjectSchema>? objects) => objects is null ? null : new Builder(objects).Build();
 
         /// <summary>
         /// Constructs a <see cref="RealmSchema"/> from an array of <see cref="Type"/> instances.
@@ -217,7 +217,7 @@ namespace Realms.Schema
         /// </returns>
         /// <seealso cref="Builder.Add(Type)"/>
         [return: NotNullIfNotNull("objects")]
-        public static implicit operator RealmSchema?(Type[]? objects) => objects == null ? null : new Builder(objects).Build();
+        public static implicit operator RealmSchema?(Type[]? objects) => objects is null ? null : new Builder(objects).Build();
 
         /// <summary>
         /// Constructs a <see cref="RealmSchema"/> from a List of <see cref="Type"/> instances.
@@ -228,7 +228,7 @@ namespace Realms.Schema
         /// </returns>
         /// <seealso cref="Builder.Add(Type)"/>
         [return: NotNullIfNotNull("objects")]
-        public static implicit operator RealmSchema?(List<Type>? objects) => objects == null ? null : new Builder(objects).Build();
+        public static implicit operator RealmSchema?(List<Type>? objects) => objects is null ? null : new Builder(objects).Build();
 
         /// <summary>
         /// Constructs a <see cref="RealmSchema"/> from a HashSet of <see cref="Type"/> instances.
@@ -239,7 +239,7 @@ namespace Realms.Schema
         /// </returns>
         /// <seealso cref="Builder.Add(Type)"/>
         [return: NotNullIfNotNull("objects")]
-        public static implicit operator RealmSchema?(HashSet<Type>? objects) => objects == null ? null : new Builder(objects).Build();
+        public static implicit operator RealmSchema?(HashSet<Type>? objects) => objects is null ? null : new Builder(objects).Build();
 
         /// <summary>
         /// A convenience operator to construct a <see cref="RealmSchema"/> from a <see cref="Builder"/> by calling the