+++ wrappers/src/app_cs.cpp
@@ -87,6 +87,8 @@ namespace realm {
             uint64_t sync_pong_keep_alive_timeout_ms;
 
             uint64_t sync_fast_reconnect_limit;
+
+            bool use_cache;
         };
     }
 }
@@ -166,7 +168,11 @@ extern "C" {
                 sync_client_config.custom_encryption_key = std::vector<char>(key.begin(), key.end());
             }
 
-            return new SharedApp(App::get_shared_app(std::move(config), std::move(sync_client_config)));
+            SharedApp app = app_config.use_cache
+                ? App::get_shared_app(std::move(config), std::move(sync_client_config))
+                : App::get_uncached_app(std::move(config), std::move(sync_client_config));
+
+            return new SharedApp(app);
         });
     }
 
@@ -296,6 +302,35 @@ extern "C" {
         }
 
         App::clear_cached_apps();
+    }
+
+    REALM_EXPORT size_t shared_app_get_base_file_path(SharedApp& app, uint16_t* buffer, size_t buffer_length, NativeException::Marshallable& ex)
+    {
+        return handle_errors(ex, [&]() {
+            std::string base_file_path(app->sync_manager()->config().base_file_path);
+            return stringdata_to_csharpstringbuffer(base_file_path, buffer, buffer_length);
+        });
+    }
+
+    REALM_EXPORT realm_string_t shared_app_get_base_uri(SharedApp& app, NativeException::Marshallable& ex)
+    {
+        return handle_errors(ex, [&]() {
+            return to_capi(app->base_url());
+        });
+    }
+
+    REALM_EXPORT realm_string_t shared_app_get_id(SharedApp& app, NativeException::Marshallable& ex)
+    {
+        return handle_errors(ex, [&]() {
+            return to_capi(app->config().app_id);
+        });
+    }
+
+    REALM_EXPORT bool shared_app_is_same_instance(SharedApp& lhs, SharedApp& rhs, NativeException::Marshallable& ex)
+    {
+        return handle_errors(ex, [&]() {
+            return lhs == rhs;  // just compare raw pointers inside the smart pointers
+        });
     }
 
 #pragma region EmailPassword